Added a jwtCarousel directive for view image galleries. #31.
The carousel is very rough at this moment. I am having issues with css.
However user is able to change images by clicking on an image in the control.
Also the images at this moment in time are hardcoded. Need to give directive
an isolated scope so that it is configurable.