Refactored upload page into a button #27.
The user can pass a label and a callback to the directive. The default label is 'Upload'.
The callback function should be the actual function that does the upload. The files
attribute of the file input is passed as the argument. When dragging and dropping
as with task #32 all the files will be passed back