Did some refactoring #27, #31 and #32.
I had to fix the upload button unit test. Found a better way of implementing the
directive. Also the upload panel passes a setter object back to the upload service
so more information can be passed. Before I was hardcoding the car id, now should
be able to upload images to any car.